+++ src/pyplier/PLIERRes.py
@@ -5,6 +5,7 @@ from pathlib import Path
 from typing import Dict, List
 
 import pandas as pd
+import numpy as np
 
 
 class PLIERResults(object):
@@ -58,6 +59,52 @@ class PLIERResults(object):
     def __str__(self) -> str:
         return self.__repr__()
 
+    def __eq__(self, other) -> bool:
+        if (
+            np.isclose(self.B, other.B).all()
+            and np.isclose(self.Z, other.Z).all()
+            and np.isclose(self.U, other.U).all()
+            and np.isclose(self.C, other.C).all()
+            and self.L1 == other.L1
+            and self.L2 == other.L2
+            and self.L3 == other.L3
+            and self.heldOutGenes == other.heldOutGenes
+            and self.withPrior == other.withPrior
+            and np.isclose(self.Uauc, other.Uauc).all()
+            and np.isclose(self.Up, other.Up).all()
+            and np.isclose(self.summary, other.summary).all()
+            and np.isclose(self.residual, other.residual).all()
+        ):
+            return True
+        else:
+            if not np.isclose(self.B, other.B).all():
+                print("B is unequal")
+            elif not np.isclose(self.Z, other.Z).all():
+                print("Z is unequal")
+            elif not np.isclose(self.U, other.U).all():
+                print("U is unequal")
+            elif not np.isclose(self.C, other.C).all():
+                print("C is unequal")
+            elif self.L1 != other.L1:
+                print("L1 is unequal")
+            elif self.L2 != other.L2:
+                print("L2 is unequal")
+            elif self.L3 != other.L3:
+                print("L3 is unequal")
+            elif self.heldOutGenes != other.heldOutGenes:
+                print("heldOutGenes is unequal")
+            elif self.withPrior != other.withPrior:
+                print("withPrior is unequal")
+            elif not np.isclose(self.Uauc, other.Uauc).all():
+                print("Uauc is unequal")
+            elif not np.isclose(self.Up, other.Up).all():
+                print("Up is unequal")
+            elif not np.isclose(self.summary, other.summary).all():
+                print("summary is unequal")
+            elif not np.isclose(self.residual, other.residual).all():
+                print("residual is unequal")
+            return False
+
     def to_dict(self):
         return {
             "B": self.B.to_dict(),
@@ -75,37 +122,6 @@ class PLIERResults(object):
             "summary": self.summary.to_dict(),
         }
 
-    def from_dict(self, source):
-        if "B" in source:
-            self.B = pd.DataFrame.from_dict(source["B"])
-        if "Z" in source:
-            self.Z = pd.DataFrame.from_dict(source["Z"])
-        if "U" in source:
-            self.U = pd.DataFrame.from_dict(source["U"])
-        if "C" in source:
-            self.C = pd.DataFrame.from_dict(source["C"])
-
-        if "L1" in source:
-            self.L1 = source["L1"]
-        if "L2" in source:
-            self.L2 = source["L2"]
-        if "L3" in source:
-            self.L3 = source["L3"]
-
-        if "heldOutGenes" in source:
-            self.heldOutGenes = source["heldOutGenes"]
-        if "withPrior" in source:
-            self.withPrior = source["withPrior"]
-
-        if "residual" in source:
-            self.residual = pd.DataFrame.from_dict(source["residual"])
-        if "Uauc" in source:
-            self.Uauc = pd.DataFrame.from_dict(source["Uauc"])
-        if "Up" in source:
-            self.Up = pd.DataFrame.from_dict(source["Up"])
-        if "summary" in source:
-            self.summary = pd.DataFrame.from_dict(source["summary"])
-
     def to_disk(self, loc: Path, compress: bool = False) -> bool:
         if not isinstance(loc, Path):
             loc = Path(loc)
@@ -118,13 +134,51 @@ class PLIERResults(object):
         return True
 
     @classmethod
+    def from_dict(cls, source):
+        pr = cls()
+        if "B" in source:
+            pr.B = pd.DataFrame.from_dict(source["B"])
+        if "Z" in source:
+            pr.Z = pd.DataFrame.from_dict(source["Z"])
+        if "U" in source:
+            pr.U = pd.DataFrame.from_dict(source["U"])
+        if "C" in source:
+            pr.C = pd.DataFrame.from_dict(source["C"])
+
+        if "L1" in source:
+            pr.L1 = source["L1"]
+        if "L2" in source:
+            pr.L2 = source["L2"]
+        if "L3" in source:
+            pr.L3 = source["L3"]
+
+        if "heldOutGenes" in source:
+            pr.heldOutGenes = source["heldOutGenes"]
+        if "withPrior" in source:
+            pr.withPrior = source["withPrior"]
+
+        if "residual" in source:
+            pr.residual = pd.DataFrame.from_dict(source["residual"])
+        if "Uauc" in source:
+            pr.Uauc = pd.DataFrame.from_dict(source["Uauc"])
+        if "Up" in source:
+            pr.Up = pd.DataFrame.from_dict(source["Up"])
+        if "summary" in source:
+            pr.summary = pd.DataFrame.from_dict(source["summary"])
+        return pr
+
+    @classmethod
     def from_disk(cls, loc: Path):
-        try:
-            with open(loc, "r") as infile:
-                input_dict = json.load(fp=infile)
-        except UnicodeDecodeError:
+        """TODO: switch to using something like HDF5 or parquet for storage"""
+        if str(loc).endswith(".gz"):
             with gzip.open(loc, "rt", encoding="UTF-8") as infile:
                 input_dict = json.load(fp=infile)
-        pr = cls()
-        pr.from_dict(input_dict)
+        else:
+            try:
+                with open(loc, "r") as infile:
+                    input_dict = json.load(fp=infile)
+            except UnicodeDecodeError:
+                with gzip.open(loc, "rt", encoding="UTF-8") as infile:
+                    input_dict = json.load(fp=infile)
+        pr = cls().from_dict(input_dict)
         return pr