+++ src/pyplier/logger.py
@@ -31,6 +31,7 @@
 
 
 import logging
+
 from typing import Optional
 
 import coloredlogs